added dynamic classes based on susuwatarii prop
