+++ src/components/r101/AppProcessStep.module.scss
@@ -85,3 +85,6 @@
     background-color: white;
   }
 }
+
+.stepContainer {
+}
